requirements.txt
pluggy
setuptools
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
